more efficient preparation - skip file if already exists

+++ devel/create_dicomdir_pilsen_pigs.py
@@ -68,14 +68,18 @@ print(df)
 for i, row in df.iterrows():
     fn_in = base_path / row["dirname"]
     fn_out = output_path / row["dataset_type"] / f"PP_{row['id']:04}"/"PATIENT_DICOM"/f"PP_{row['id']:04}.mhd"
+    fn_meta = fn_out.parent / "meta.json"
     logger.debug(fn_in)
     logger.debug(fn_out)
 
+
     fn_out.parent.mkdir(parents=True, exist_ok=True)
-    datap = io3d.read(fn_in)
-    io3d.write(datap, fn_out)
-    with open(fn_out.parent / "meta.json", "w") as f:
-        json.dump(dict(row), f)
+    if not fn_meta.exists(): # we do not need to read the data again if everything is done. We are checkin:w
+
+        datap = io3d.read(fn_in)
+        io3d.write(datap, fn_out)
+        with open(fn_meta, "w") as f:
+            json.dump(dict(row), f)
